Lots of updates for Message Tester
- Much progress on converting the page to bootstrap
- convert storage API to using a database model instead of Django cache
- add buttons to clear messages related to a particular identity, or
  all the messages
- add a new doc page, with synopsis of what it's for and how to use it,
  and install instructions
- add tests, verify 100% coverage of the httptester/*.py files
- check and fix most PEP-8 violations

Ticket: #215
Branch: feature/httptester-update

+++ rapidsms/templatetags/paginator_tags.py
@@ -33,8 +33,8 @@ def paginator(objects):
     # TODO: gah, extract this junk a private method
     max_page_links = settings.PAGINATOR_MAX_PAGE_LINKS
     last_page_number = objects.paginator.num_pages + 1
-    last_low_number = math.floor(max_page_links / 2)
-    first_high_number = last_page_number - math.ceil(max_page_links / 2)
+    last_low_number = int(math.floor(max_page_links / 2))
+    first_high_number = last_page_number - int(math.ceil(max_page_links / 2))
 
     page_links = [
         _page(number)
